+++ airbyte-cdk/python/unit_tests/sources/streams/concurrent/test_concurrent_read_processor.py
@@ -40,13 +40,11 @@ class TestConcurrentReadProcessor(unittest.TestCase):
         self._thread_pool_manager = Mock(spec=ThreadPoolManager)
 
         self._an_open_partition = Mock(spec=Partition)
-        self._an_open_partition.is_closed.return_value = False
         self._log_message = Mock(spec=LogMessage)
         self._an_open_partition.to_slice.return_value = self._log_message
         self._an_open_partition.stream_name.return_value = _STREAM_NAME
 
         self._a_closed_partition = Mock(spec=Partition)
-        self._a_closed_partition.is_closed.return_value = True
         self._a_closed_partition.stream_name.return_value = _ANOTHER_STREAM_NAME
 
         self._logger = Mock(spec=logging.Logger)
@@ -75,6 +73,19 @@ class TestConcurrentReadProcessor(unittest.TestCase):
         self._record = Mock(spec=Record)
         self._record.stream_name = _STREAM_NAME
         self._record.data = self._record_data
+
+    def test_stream_is_not_done_initially(self):
+        stream_instances_to_read_from = [self._stream]
+        handler = ConcurrentReadProcessor(
+            stream_instances_to_read_from,
+            self._partition_enqueuer,
+            self._thread_pool_manager,
+            self._logger,
+            self._slice_logger,
+            self._message_repository,
+            self._partition_reader,
+        )
+        assert not handler._is_stream_done(self._stream.name)
 
     def test_handle_partition_done_no_other_streams_to_generate_partitions_for(self):
         stream_instances_to_read_from = [self._stream]
@@ -111,7 +122,6 @@ class TestConcurrentReadProcessor(unittest.TestCase):
             self._partition_reader,
         )
         handler.start_next_partition_generator()
-        handler.on_partition(self._a_closed_partition)
 
         sentinel = PartitionGenerationCompletedSentinel(self._another_stream)
         messages = handler.on_partition_generation_completed(sentinel)
@@ -147,7 +157,7 @@ class TestConcurrentReadProcessor(unittest.TestCase):
         handler.on_partition(self._a_closed_partition)
 
         self._thread_pool_manager.submit.assert_called_with(self._partition_reader.process_partition, self._a_closed_partition)
-        assert self._a_closed_partition in handler._streams_to_partitions[_ANOTHER_STREAM_NAME]
+        assert self._a_closed_partition in handler._streams_to_running_partitions[_ANOTHER_STREAM_NAME]
 
     def test_handle_partition_emits_log_message_if_it_should_be_logged(self):
         stream_instances_to_read_from = [self._stream]
@@ -169,15 +179,16 @@ class TestConcurrentReadProcessor(unittest.TestCase):
 
         self._thread_pool_manager.submit.assert_called_with(self._partition_reader.process_partition, self._an_open_partition)
         self._message_repository.emit_message.assert_called_with(self._log_message)
-        assert self._an_open_partition in handler._streams_to_partitions[_STREAM_NAME]
 
+        assert self._an_open_partition in handler._streams_to_running_partitions[_STREAM_NAME]
+
+    @freezegun.freeze_time("2020-01-01T00:00:00")
     def test_handle_on_partition_complete_sentinel_with_messages_from_repository(self):
         stream_instances_to_read_from = [self._stream]
         partition = Mock(spec=Partition)
         log_message = Mock(spec=LogMessage)
         partition.to_slice.return_value = log_message
         partition.stream_name.return_value = _STREAM_NAME
-        partition.is_closed.return_value = True
 
         handler = ConcurrentReadProcessor(
             stream_instances_to_read_from,
@@ -189,6 +200,7 @@ class TestConcurrentReadProcessor(unittest.TestCase):
             self._partition_reader,
         )
         handler.start_next_partition_generator()
+        handler.on_partition(partition)
 
         sentinel = PartitionCompleteSentinel(partition)
 
@@ -223,6 +235,7 @@ class TestConcurrentReadProcessor(unittest.TestCase):
             self._partition_reader,
         )
         handler.start_next_partition_generator()
+        handler.on_partition(self._a_closed_partition)
         handler.on_partition_generation_completed(PartitionGenerationCompletedSentinel(self._another_stream))
 
         sentinel = PartitionCompleteSentinel(self._a_closed_partition)
@@ -254,7 +267,6 @@ class TestConcurrentReadProcessor(unittest.TestCase):
         log_message = Mock(spec=LogMessage)
         partition.to_slice.return_value = log_message
         partition.stream_name.return_value = _STREAM_NAME
-        partition.is_closed.return_value = True
 
         handler = ConcurrentReadProcessor(
             stream_instances_to_read_from,
@@ -282,7 +294,6 @@ class TestConcurrentReadProcessor(unittest.TestCase):
         log_message = Mock(spec=LogMessage)
         partition.to_slice.return_value = log_message
         partition.stream_name.return_value = _STREAM_NAME
-        partition.is_closed.return_value = True
         self._message_repository.consume_queue.return_value = []
 
         handler = ConcurrentReadProcessor(
@@ -319,7 +330,6 @@ class TestConcurrentReadProcessor(unittest.TestCase):
         log_message = Mock(spec=LogMessage)
         partition.to_slice.return_value = log_message
         partition.stream_name.return_value = _STREAM_NAME
-        partition.is_closed.return_value = True
         slice_logger = Mock(spec=SliceLogger)
         slice_logger.should_log_slice_message.return_value = True
         slice_logger.create_slice_log_message.return_value = log_message
@@ -370,7 +380,6 @@ class TestConcurrentReadProcessor(unittest.TestCase):
         stream_instances_to_read_from = [self._stream]
         partition = Mock(spec=Partition)
         partition.stream_name.return_value = _STREAM_NAME
-        partition.is_closed.return_value = True
 
         handler = ConcurrentReadProcessor(
             stream_instances_to_read_from,
@@ -413,7 +422,6 @@ class TestConcurrentReadProcessor(unittest.TestCase):
         log_message = Mock(spec=LogMessage)
         partition.to_slice.return_value = log_message
         partition.stream_name.return_value = _STREAM_NAME
-        partition.is_closed.return_value = True
         self._message_repository.consume_queue.return_value = [
             AirbyteMessage(type=MessageType.LOG, log=AirbyteLogMessage(level=LogLevel.INFO, message="message emitted from the repository"))
         ]
@@ -474,7 +482,69 @@ class TestConcurrentReadProcessor(unittest.TestCase):
             self._message_repository,
             self._partition_reader,
         )
-        handler._streams_to_partitions = {_STREAM_NAME: {self._an_open_partition}, _ANOTHER_STREAM_NAME: {self._a_closed_partition}}
+
+        handler.start_next_partition_generator()
+
+        another_stream = Mock(spec=AbstractStream)
+        another_stream.name = _STREAM_NAME
+        another_stream.as_airbyte_stream.return_value = AirbyteStream(
+            name=_ANOTHER_STREAM_NAME,
+            json_schema={},
+            supported_sync_modes=[SyncMode.full_refresh],
+        )
+
+        exception = RuntimeError("Something went wrong")
+
+        messages = []
+
+        with self.assertRaises(RuntimeError):
+            for m in handler.on_exception(exception):
+                messages.append(m)
+
+        expected_message = [
+            AirbyteMessage(
+                type=MessageType.TRACE,
+                trace=AirbyteTraceMessage(
+                    type=TraceType.STREAM_STATUS,
+                    emitted_at=1577836800000.0,
+                    stream_status=AirbyteStreamStatusTraceMessage(
+                        stream_descriptor=StreamDescriptor(name=_STREAM_NAME), status=AirbyteStreamStatus(AirbyteStreamStatus.INCOMPLETE)
+                    ),
+                ),
+            ),
+            AirbyteMessage(
+                type=MessageType.TRACE,
+                trace=AirbyteTraceMessage(
+                    type=TraceType.STREAM_STATUS,
+                    emitted_at=1577836800000.0,
+                    stream_status=AirbyteStreamStatusTraceMessage(
+                        stream_descriptor=StreamDescriptor(name=_ANOTHER_STREAM_NAME), status=AirbyteStreamStatus(AirbyteStreamStatus.INCOMPLETE)
+                    ),
+                ),
+            )
+        ]
+
+        assert messages == expected_message
+        self._thread_pool_manager.shutdown.assert_called_once()
+
+    @freezegun.freeze_time("2020-01-01T00:00:00")
+    def test_on_exception_does_not_stop_streams_that_are_already_done(self):
+        stream_instances_to_read_from = [self._stream, self._another_stream]
+
+        handler = ConcurrentReadProcessor(
+            stream_instances_to_read_from,
+            self._partition_enqueuer,
+            self._thread_pool_manager,
+            self._logger,
+            self._slice_logger,
+            self._message_repository,
+            self._partition_reader,
+        )
+
+        handler.start_next_partition_generator()
+        handler.on_partition(self._an_open_partition)
+        handler.on_partition_generation_completed(PartitionGenerationCompletedSentinel(self._stream))
+        handler.on_partition_generation_completed(PartitionGenerationCompletedSentinel(self._another_stream))
 
         another_stream = Mock(spec=AbstractStream)
         another_stream.name = _STREAM_NAME