+++ airbyte-cdk/python/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.59.1
+Fix concurrent CDK deadlock
+
+## 0.59.0
+Fix state message handling when running concurrent syncs
+
+## 0.58.9
+concurrent-cdk: improve resource usage when reading from substreams
+
 ## 0.58.8
 CDK: HttpRequester can accept http_method in str format, which is required by custom low code components