+++ airbyte-integrations/bases/connector-acceptance-test/CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
-## 2.3.0
+## 3.2.0
+Add TestBasicRead.test_all_supported_file_types_present, which validates that all supported file types are present in the sandbox account for certified file-based connectors.
+
+## 3.1.0
+Add TestSpec.test_oauth_is_default_method test with OAuth is default option validation.
+
+## 3.0.1
+Upgrade to Dagger 0.9.6
+
+## 3.0.0
 Upgrade to Dagger 0.9.5
 
 ## 2.2.0